pull

+++ src/main/java/com/effourt/calenkit/controller/MemberController2.java
@@ -6,23 +6,16 @@ import com.effourt.calenkit.repository.MemberRepository;
 import com.effourt.calenkit.service.AdminService;
 import com.effourt.calenkit.service.MyPageService;
 import com.effourt.calenkit.util.ImageUpload;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.util.Collections;
 import java.util.List;
 
 @Slf4j
@@ -39,7 +32,7 @@ public class MemberController2 {
     /** 관리자 */
     // Admin
     @GetMapping(value = "/admin")
-    public String AdminList(@RequestParam(required = false) String keyword, Model model) {
+    public String Admin(@RequestParam(required = false) String keyword, Model model) {
         model.addAttribute("keyword", keyword);
         return "member/admin";
     }
@@ -58,12 +51,12 @@ public class MemberController2 {
     }
 
 
-    // Admin
-    // 멤버 상태변경(PATCH)
-    //  member_list 페이지로 이동
-    @PostMapping(value ="/admin_statusModify")
-    public String AdminModify(@ModelAttribute Member member) throws MemberNotFoundException {
-        adminService.modifyMember(member);
+
+    @GetMapping(value = "/admin_modifyMember")
+    public String AdminModify(@RequestParam("selectedValue") Integer memStatus, String memId) {
+        Member member = memberRepository.findByMemId(memId);
+        member.setMemStatus(memStatus);
+        adminService.modifyStatus(member);
         return "member/admin";
     }
 
@@ -71,14 +64,20 @@ public class MemberController2 {
     // 멤버 삭제(DELETE)
     // 로그인세션에서 아이디값을 전달받아 member_list 페이지로 이동
     @PostMapping(value ="/admin_delete")
-    public String adminDelete(List<String> memIdList) throws MemberNotFoundException {
-       //체크 된 멤버리스트 삭제하기 위해 for문 사용.
+    public String adminDelete(@RequestParam("memIdList") List<String> memIdList) throws MemberNotFoundException {
+        //체크 된 멤버리스트 삭제하기 위해 for문 사용.
         for (String originMemId : memIdList) {
             //리스트로 객체 받아올 경우 [,],"가 포함되어있는데 이걸 제거하기 위해 사용함.
             String originalString =originMemId.toString();
             String memId = originalString.replaceAll("[\\[\\]\",]", "");
-            adminService.removeMember(memId);
-        }
+            //본인 계정 삭제 불가능
+            if(memId.equals(session.getAttribute("loginId"))){
+                return "member/admin";
+            }
+            else {
+                adminService.removeMember(memId);
+            }
+       }
         return "member/admin";
     }
 
@@ -100,16 +99,14 @@ public class MemberController2 {
     @GetMapping("/nameCheck")
     @ResponseBody
     public int nameCheck(String memName) {
-
-        if (!memName.matches("^[a-zA-Z가-힣]{2,10}$")) {
-            int cnt = memberRepository.findByMemName(memName);
-            return cnt;
-        }
-        else{
-            int cnt=2;
-            session.setAttribute("errorMessage", "닉네임은 2~10자의 한글 또는 영문으로 입력해주세요.");
-            return cnt;
-        }
+        int cnt=0;
+            if (memName.matches("^[a-zA-Z가-힣0-9]{2,10}$")) {
+                cnt = memberRepository.findByMemName(memName);
+                return cnt;
+            } else {
+                cnt = 2;
+                return cnt;
+            }
     }
     // MyPage
     // 아이디 검색 후 중복 확인(GET)
@@ -129,38 +126,9 @@ public class MemberController2 {
         return cnt;
     }
 
-    // MyPage
-    // 아이디 검색 후 중복 확인(GET)
-    // Ajax 처리를 위해 비밀번호 중복 갯수 반환
-    @GetMapping("/passwordCheck")
-    @ResponseBody
-    public int passwordCheck(String password1, String password2) {
-        int cnt = 0;
-        if (password1.equals(password2)) {
-            cnt++;
-        } else if (password1.matches("^(?=.*\\d)(?=.*[a-zA-Z])(?=.*[!@#$%^&+=])(?=^.{8,15}$)")) {
-            cnt = 2;
-        }
-        return cnt;
-    }
-
-    @GetMapping("/pwCheck")
-    @ResponseBody
-    public int pwCheck(String memPw) {
-        String loginId=(String)session.getAttribute("loginId");
-        Member loginMember=memberRepository.findByMemId(loginId);
-        int cnt=0;
-        if(memPw==null){
-        }
-        if (passwordEncoder.matches(memPw,loginMember.getMemPw())){
-          cnt++;
-            return cnt;
-        }
-        return cnt;
-    }
 
     @PostMapping("/modify_image")
-    public String saveMember(@RequestParam(required = false ) MultipartFile memImage) throws MemberNotFoundException, IOException {
+    public String saveImage(@RequestParam(required = false ) MultipartFile memImage) throws MemberNotFoundException, IOException {
         String loginId=(String)session.getAttribute("loginId");
         Member loginMember=memberRepository.findByMemId(loginId);
         // 이미지 업로드 후 파일명 반환
@@ -173,8 +141,11 @@ public class MemberController2 {
         return "redirect:/myPage";
     }
 
+    // MyPage
+    // 닉네임 검색 후 중복 확인(GET)
+    // 형식에 맞거나 사용하지 않는 닉네임 Ajax처리.
     @PostMapping("/modify_name")
-    public String Member(String memName) throws MemberNotFoundException, IOException {
+    public String saveName(String memName) throws MemberNotFoundException, IOException {
         String loginId=(String)session.getAttribute("loginId");
         Member loginMember=memberRepository.findByMemId(loginId);
         // 닉네임 업로드
@@ -186,12 +157,63 @@ public class MemberController2 {
 
 
 
-    @GetMapping(value ="/myPage_pwModify")
-    public String MyPagePwModify(Model model){
+    // MyPage
+    // 아이디 검색 후 중복 확인(GET)
+    // 로그인 계정 비밀번호와 비밀번호 확인 Ajax 비교
+    @GetMapping("/pwCheck")
+    @ResponseBody
+    public int pwCheck(String memPw) {
         String loginId=(String)session.getAttribute("loginId");
         Member loginMember=memberRepository.findByMemId(loginId);
+        int cnt=0;
+        if(memPw==null){
+        }
+        if (passwordEncoder.matches(memPw,loginMember.getMemPw())){
+            cnt++;
+            return cnt;
+        }
+        return cnt;
+    }
 
-        model.addAttribute("loginMember"+loginMember);
+
+    // MyPage
+    // 아이디 검색 후 중복 확인(GET)
+    // Ajax 처리를 위해 비밀번호 중복 갯수 반환
+    @GetMapping("/passwordCheck")
+    @ResponseBody
+    public int passwordCheck(String password1, String password2) {
+        int cnt = 0;
+        if(password1.matches("(?=.*\\d)(?=.*[a-z])(?=.*[!@#])[\\da-zA-Z!@#]{8,15}")) {
+            System.out.println("cnt1="+cnt);
+            if (password2.equals(password1)) {
+                cnt++; //
+                System.out.println("cnt2="+cnt);
+                return cnt; //1 출력 password2와 password1이 일치할 경우
+            }
+        }
+        return cnt; //0 출력
+    }
+    @GetMapping("/passwordCheck2")
+    @ResponseBody
+    public int passwordCheck2(String password1) {
+        int cnt = 0;
+        if(password1.matches("(?=.*\\d)(?=.*[a-z])(?=.*[!@#])[\\da-zA-Z!@#]{8,15}")) {
+            System.out.println("cnt1="+cnt);
+                cnt++;
+                System.out.println("cnt2="+cnt);
+                return cnt; //1 출력 password2와 password1이 일치할 경우
+        }
+        return cnt; //0 출력
+    }
+    // MyPage
+    // 멤버 비밀번호 정보변경(Get)
+    @GetMapping(value ="/myPage_pwModify")
+    public String savePw(Model model){
+        String loginId=(String)session.getAttribute("loginId");
+        Member loginMember=memberRepository.findByMemId(loginId);
+        System.out.print("loginId="+loginId);
+        System.out.print("loginId="+loginMember.getMemPw());
+        model.addAttribute("loginMember",loginMember);
         return "member/myPageModify";
     }
